Add Reading
As it turns out the more research I keep doing on the game the longer the results keep getting. Freeciv has undoubtedly stood the test of time and is well known if not immensely popular in the RL community. This thing is only making me more confident in the project.